src/main/java/wniemiec/mobilex/ama/parser/screens/behavior/instruction/ArrayPatternInstructionJsonParser.java
package wniemiec.mobilex.ama.parser.screens.behavior.instruction;

import org.json.JSONException;
import org.json.JSONObject;
import wniemiec.mobilex.ama.models.behavior.ArrayPattern;
import wniemiec.mobilex.ama.models.behavior.Instruction;
import wniemiec.mobilex.ama.parser.exception.ParseException;
import wniemiec.mobilex.ama.parser.screens.behavior.expression.ExpressionParser;


/**
 * Responsible for parsing array pattern statements from behavior node from 
 * MobiLang AST.
 */
class ArrayPatternInstructionJsonParser implements InstructionJsonParser {
    
    //-------------------------------------------------------------------------
    //		Attributes
    //-------------------------------------------------------------------------
    private static ArrayPatternInstructionJsonParser instance;
    private final ExpressionParser expressionParser;


    //-------------------------------------------------------------------------
    //		Constructor
    //-------------------------------------------------------------------------
    private ArrayPatternInstructionJsonParser() {
        expressionParser = ExpressionParser.getInstance();
    }
    


    //-------------------------------------------------------------------------
    //		Factory
    //-------------------------------------------------------------------------
    public static ArrayPatternInstructionJsonParser getInstance() {
        if (instance == null) {
            instance = new ArrayPatternInstructionJsonParser();
        }

        return instance;
    }
    

    //-------------------------------------------------------------------------
    //		Methods
    //-------------------------------------------------------------------------
    @Override
    public Instruction parse(JSONObject jsonObject) 
    throws JSONException, ParseException {
        return new ArrayPattern(
            expressionParser.parse(jsonObject.getJSONArray("elements"))
        );
    }
}
